requirements.txt
playwright
twocaptcha-python
